feat(lexer): Initial lexer and token framework
This commit lays the framework for the lexer, token, and repl.

The initial tokens defined are single character terminals such as
delimiters, parenthesis, braces, etc.

The initial lexer implementation includes support for creating a lexer,
parsing the input to create tokens, and generating the tokens.

Units for the lexer cover:

    - Creating a lexer
    - Reading characters
    - Creating tokens

The repl package is just an empty package at the moment.