feat(lexer): add support for identifiers and keywords
Two new functions were introduced; isLetter and readIdentifier.  These
functions are responsible for reading a sequence of characters that are
part of the legal name for an identifier or keyword and making the
distinction between the two.

A keyword always has higher precedence over an identifier.  Meaning the
occurrence of a sequence of letters that matches an entry in the keyword
map will default to that keyword and not an identifier.