feat(parser): add initial error message support
This commit adds the ability to capture errors and report them while
parsing the program.

This minimal support is expected to be extended and changed as
additional language support is added to the parser.

The TestLetStatementErrors unit test is minimal in that it only tests
that the number of errors reported matches what is expected.  The actual
errors are not validated as part of the test.

This will be developed as error handling matures.