feat(parser): add support for prefix expression parsing
Support for parsing expressions with the `!` and `-` prefix operators
can now be handled.

It's important to understand the protocol used.

`parsePrefixExpression` starts with `p.curToken` being the token of the
prefix operator and it returns with p.curToken being the operand of the
prefix expression, which is the last token of the expression.