refactor(module): reconfigure module and packages to point to github (#10)
Prior to this refactor, the project relied on manipulating the `GOPATH`
environment variable to point to the parent directory of `src` (i.e. the
root of the repository).  While this approach works fine, it's not the
idiomatic way of organizing a Go module.

This commit reorganizes the module layout to conform with Go's best
practices for module and package organization.