refactor(tests): refactor all unit tests
With support for expression parsing making progress, this is a good spot
to refactor the tests with helper functions for easier development of
additional tests.

Additionally, with enough expression support already developed, the let
and return statement parser functions will now properly parse the
expression portion of the test and we do not need to skip over the
expression.