feat(env): add a runtime environment for let statement support (#40)
The runtime environment enables the interpreter to persist state for
variables.  The initial implementation is a hash map of strings to
objects that store the identifier as the key and the object as the
value.  This will be extended when function call handling is added.

As part of the testing and support for the environment, let statement
functionality is also part of this commit.